requirements.txt
openai
python-dotenv 
openai-whisper
youtube_dl
fastapi
sounddevice
wavio
langchain-openai
langchain-pinecone
langchain-community
groq
streamlit 
faster-whisper
faster-whisper
sounddevice
numpy
wavio
langchain-openai
langchain-core
langchain-community
torch
streamlit
yt-dlp